Rework the report querying.
Pass contexts in instead of creating them in the database code.
